Refactor instance pipeline into modules
